+++ CHANGELOG.md
@@ -418,3 +418,4 @@
 
 # 0.4.1
 * Generalize `$` and `#` `Category`'s infix operators
+* Add `Covariant` constraint in `Bivariant` typeclass definition